requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
streamlit
streamlit-toggle